Refactor to define core privacy pass protocol.
