+++ test/authScheme.test.ts
@@ -1,8 +1,8 @@
 // Copyright (c) 2023 Cloudflare, Inc.
 // Licensed under the Apache-2.0 license found in the LICENSE file or at https://opensource.org/licenses/Apache-2.0
 
-import { TOKEN_TYPES } from '../src/pubVerifToken.js';
-import { TokenChallenge, TokenPayload, PrivateToken } from '../src/httpAuthScheme.js';
+import { TOKEN_TYPES, WWWAuthenticateHeader } from '../src/index.js';
+import { TokenChallenge, TokenPayload } from '../src/tokenBase.js';
 import { hexToString, hexToUint8, testSerialize, uint8ToHex } from './util.js';
 
 // Test vectors generated by pat-go (https://github.com/cloudflare/pat-go/)
@@ -32,7 +32,13 @@ test.each(tokenVectors)('AuthScheme-TokenVector-%#', async (v: TokenVectors) => 
     testSerialize(TokenChallenge, challenge);
 
     const context = new Uint8Array(await crypto.subtle.digest('SHA-256', challengeSerialized));
-    const payload = new TokenPayload(TOKEN_TYPES.BLIND_RSA, nonce, context, keyId);
+    const payload = new TokenPayload(
+        TOKEN_TYPES.BLIND_RSA,
+        TOKEN_TYPES.BLIND_RSA.value,
+        nonce,
+        context,
+        keyId,
+    );
     const payloadEnc = payload.serialize();
 
     expect(uint8ToHex(payloadEnc)).toBe(v.token_authenticator_input);
@@ -41,11 +47,10 @@ test.each(tokenVectors)('AuthScheme-TokenVector-%#', async (v: TokenVectors) => 
 type HeaderVectors = (typeof headerVectors)[number];
 
 test.each(headerVectors)('AuthScheme-HeaderVector-%#', (v: HeaderVectors) => {
-    const tokens = PrivateToken.parse(v['WWW-Authenticate']);
+    const tokens = WWWAuthenticateHeader.parse(v['WWW-Authenticate']);
 
     let i = 0;
     for (const t of tokens) {
-        expect(uint8ToHex(t.challengeSerialized)).toBe(v[`token-challenge-${i}` as keyof typeof v]);
         expect(uint8ToHex(t.tokenKey)).toBe(v[`token-key-${i}` as keyof typeof v]);
         expect(t.challenge.tokenType).toBe(v[`token-type-${i}` as keyof typeof v]);
         expect(t.maxAge).toBe(v[`max-age-${i}` as keyof typeof v]);
